+++ renderer/tests/suite/main.rs
@@ -23,6 +23,7 @@ mod handshake;
 mod painting;
 mod pictures;
 mod portable_maths;
+mod header_queue;
 mod header_scenes;
 mod headless;
 mod headless_measures;